rosout_filter: node to filter log messages
This new node reads a specified topic (/rosout_agg by default) of type
roslib::Log or rosgraph_msgs::Log (depending on ROS version) and filters
it by a chain of conditions specified in a YAML file. The messages which
are not filtered out by any of the rules or which are explicitly
filtered for inclusion are forwarded to another topic (/rosout_filtered
by default).